+++ src/main/kotlin/racing_car/game/Application.kt
@@ -1,0 +1,7 @@
+package racing_car.game
+
+import racing_car.game.ui.CommandLineReceiver
+
+fun main() {
+    Game(CommandLineReceiver()).start()
+}
