+++ src/main/kotlin/racing_car/game/Participant.kt
@@ -1,0 +1,11 @@
+package racing_car.game
+
+class Participant(numOfCar: Int) {
+    val cars = emptyList<Car>().toMutableList()
+
+    init {
+        for (idx in 1..numOfCar) {
+            cars.add(Car())
+        }
+    }
+}
