--- src/main/kotlin/step3_racing_car/game/Application.kt
@@ -1,7 +1,0 @@
-package step3_racing_car.game
-
-import step3_racing_car.game.ui.CommandLineReceiver
-
-fun main() {
-    Game(CommandLineReceiver()).start()
-}
